wip(mongodb): phase 3.3 - thu hep loi tu 621 -> 136 dong log (chi con 7 file)
DA SUA (chu yeu bang lop tuong thich / compat adapter, khong phai viet lai service):
- CartItem: getBook()/setBook() (Book "nhe" tu snapshot)
- Coupon: getSeller()/setSeller(User) | RefreshToken: getUser()/setUser(User)
- Notification: getUser()/setUser(User) | NotificationDelivery: getNotification()/setNotification()
- Customer: getUser() | BookReview: setBook(Book)/setUser(User) (luu BookRef + UserSnapshot)
- AssociationRule: getBookA()/getBookB()/setBookA/setBookB
- Category: them constructor 4 tham so (id, name, description, parentId)
- User: getFavoriteCategories()/setFavoriteCategories(Set<Category>) (chuyen sang favoriteCategoryIds)
- BookRepository: findBySeller(User|UserSnapshot), countBooksByCategoryAndSeller(sellerId, String)
- OrderRepository: countByBuyer/sumTotalAmountByBuyer/sumDiscountAmountByBuyer/countDiscountedOrdersByBuyer/
  findLastOrderDateByBuyer/findByBuyerOrderByCreatedAtDesc/findByBuyerWithFilters (nhan entity User)
- OrderReturnRepository, SupportTicketRepository: default method nhan User
- CouponRepository: findBySeller_IdOrderByCreatedAtDesc / existsByCodeIgnoreCaseAndSeller_Id
- AssociationRuleRepository: deleteAllRules(), findBoughtTogetherByBookId(..., BigDecimal)
- OrderItem: getBook() dung BookImage (thay vi imageUrl builder)

DA SUA TRUC TIEP:
- DatabaseSeederService: Optional unwrap, buyer/user snapshot, orderId/subOrderId thay parentOrder/subOrder
- CustomerService: .userId(user.getId())
- RecommendationJob: BigDecimal -> doubleValue() cho support/confidence/lift
- BookService: goi searchApprovedBooks(...) thay searchApprovedBooksNative(...)
- BookReviewService: bookId/userId thay book/user khi tao review; Optional unwrap; generics List<Map>
- NotificationDeliveryQueue: .notificationId/.userId thay .notification(...)
- VnpayService: .orderId(order.getId())
- DistributedLockService: findAndModify thay 3 stored procedure
- MOI: ReviewSnapshotCallback (tu dong lam day snapshot review.book/review.user khi ghi)

CON LAI 7 FILE (~20 diem sua nho):
1. OrderService (90,94,108,136,499,502,859): UserSnapshot <-> User + builder parentOrder/subOrder
2. AuthService (67,113,121,151,180): seller snapshot + favoriteCategories + Optional
3. DatabaseSeederService (128,145): shippingAddress(...) va book(...) trong builder
4. NotificationService (87,167): ep int + .userId(...)
5. UserService (28,125): Optional + snapshot
6. DataSeeder (58): Optional
7. converter/AssociationRuleConverter (24): getRuleId() -> getId()
Chua tinh 15 file test (chuyen @DataJpaTest -> @DataMongoTest).

+++ src/main/java/com/example/bookstore/service/recommendation/RecommendationJob.java
@@ -110,9 +110,9 @@ public class RecommendationJob {
                 dbRule.setBookB(bookBOpt.get());
                 
                 // Set chính xác support, confidence, lift từ thuật toán
-                dbRule.setSupport(BigDecimal.valueOf(rule.getSupport()).setScale(4, java.math.RoundingMode.HALF_UP));
-                dbRule.setConfidence(BigDecimal.valueOf(rule.getConfidence()).setScale(4, java.math.RoundingMode.HALF_UP));
-                dbRule.setLift(BigDecimal.valueOf(rule.getLift()).setScale(4, java.math.RoundingMode.HALF_UP));
+                dbRule.setSupport(BigDecimal.valueOf(rule.getSupport()).setScale(4, java.math.RoundingMode.HALF_UP).doubleValue());
+                dbRule.setConfidence(BigDecimal.valueOf(rule.getConfidence()).setScale(4, java.math.RoundingMode.HALF_UP).doubleValue());
+                dbRule.setLift(BigDecimal.valueOf(rule.getLift()).setScale(4, java.math.RoundingMode.HALF_UP).doubleValue());
                 dbRule.setUpdatedAt(LocalDateTime.now());
                 
                 dbRules.add(dbRule);
